Disable gunicorn's console handler, format things.
Gunicorn is attaching a logging console handler no matter what options I
provide it in the dict in serve.py. This leaves us in a situation where if it
is enabled, you get duplicate messages from gunicorn with different formats.
The alternative is to disable our console handler and let gunicorn do its own
thing, which unfortunately means that we'll two different formats on the
console. Note that the log file (xmr.log) itself looks fine and consistent.